requirements.txt
Flask
python-dotenv
cloudinary
pycryptodome
requests
Flask-Mail
itsdangerous
gunicorn